+++ src/OpenSimCreator/UI/MeshImporter/MeshImporterSharedState.h
@@ -419,7 +419,7 @@ namespace osc::mi
             scene_renderer_.render(decs, p);
 
             // send texture to ImGui
-            ui::draw_image(scene_renderer_.upd_render_texture(), scene_renderer_.dimensions());
+            ui::draw_image(scene_renderer_.upd_render_texture());
 
             // handle hittesting, etc.
             setIsRenderHovered(ui::is_item_hovered(ImGuiHoveredFlags_AllowWhenBlockedByPopup));
@@ -506,22 +506,22 @@ namespace osc::mi
 
         DrawableThing generateFloorDrawable() const
         {
-            Transform t = calc_floor_transform();
-            t.scale *= 0.5f;
-
             auto props = MeshBasicMaterial::PropertyBlock{};
             props.set_color(m_Colors.gridLines);
 
-            DrawableThing dt;
-            dt.id = MIIDs::Empty();
-            dt.groupId = MIIDs::Empty();
-            dt.mesh = App::singleton<SceneCache>(App::resource_loader())->grid_mesh();
-            dt.transform = t;
-            dt.color = m_Colors.gridLines;
-            dt.flags = SceneDecorationFlags::None;
-            dt.material = m_FloorMaterial;
-            dt.maybePropertyBlock = props;
-            return dt;
+            return DrawableThing{
+                .id = MIIDs::Empty(),
+                .groupId = MIIDs::Empty(),
+                .mesh = App::singleton<SceneCache>(App::resource_loader())->grid_mesh(),
+                .transform = {
+                    .scale = 0.5f * Vec3{m_SceneScaleFactor * 100.0f, m_SceneScaleFactor * 100.0f, 1.0f},
+                    .rotation = angle_axis(90_deg, Vec3{-1.0f, 0.0f, 0.0f}),
+                },
+                .color = m_Colors.gridLines,
+                .flags = SceneDecorationFlag::None,
+                .material = m_FloorMaterial,
+                .maybePropertyBlock = props,
+            };
         }
 
         //
@@ -651,14 +651,13 @@ namespace osc::mi
 
         DrawableThing generateMeshDrawable(const osc::mi::Mesh& el) const
         {
-            DrawableThing rv;
-            rv.id = el.getID();
-            rv.groupId = MIIDs::MeshGroup();
-            rv.mesh = el.getMeshData();
-            rv.transform = el.getXForm();
-            rv.color = el.getParentID() == MIIDs::Ground() || el.getParentID() == MIIDs::Empty() ? RedifyColor(getColorMesh()) : getColorMesh();
-            rv.flags = SceneDecorationFlags::None;
-            return rv;
+            return DrawableThing{
+                .id = el.getID(),
+                .groupId = MIIDs::MeshGroup(),
+                .mesh = el.getMeshData(),
+                .transform = el.getXForm(),
+                .color = el.getParentID() == MIIDs::Ground() || el.getParentID() == MIIDs::Empty() ? RedifyColor(getColorMesh()) : getColorMesh(),
+            };
         }
 
         void appendDrawables(
@@ -707,7 +706,7 @@ namespace osc::mi
                         el.getXForm(),
                         appendOut,
                         1.0f,
-                        SceneDecorationFlags::None,
+                        SceneDecorationFlag::Default,
                         GetJointAxisLengths(el)
                     );
                 },
@@ -1164,14 +1163,6 @@ namespace osc::mi
             m_VisibilityFlags.stationConnectionLines = newIsShowing;
         }
 
-        Transform calc_floor_transform() const
-        {
-            return {
-                .scale = {m_SceneScaleFactor * 100.0f, m_SceneScaleFactor * 100.0f, 1.0f},
-                .rotation = angle_axis(90_deg, Vec3{-1.0f, 0.0f, 0.0f}),
-            };
-        }
-
         float getSphereRadius() const
         {
             return 0.02f * m_SceneScaleFactor;
@@ -1188,7 +1179,7 @@ namespace osc::mi
             const Transform& xform,
             std::vector<DrawableThing>& appendOut,
             float alpha = 1.0f,
-            SceneDecorationFlags flags = SceneDecorationFlags::None,
+            SceneDecorationFlags flags = SceneDecorationFlag::Default,
             Vec3 legLen = {1.0f, 1.0f, 1.0f},
             Color coreColor = Color::white()) const
         {
@@ -1260,42 +1251,37 @@ namespace osc::mi
                 Transform scaled{xform};
                 scaled.scale *= halfWidth;
 
-                DrawableThing& originCube = appendOut.emplace_back();
-                originCube.id = logicalID;
-                originCube.groupId = groupID;
-                originCube.mesh = App::singleton<SceneCache>(App::resource_loader())->brick_mesh();
-                originCube.transform = scaled;
-                originCube.color = Color::white();
-                originCube.flags = SceneDecorationFlags::None;
+                appendOut.push_back(DrawableThing{
+                    .id = logicalID,
+                    .groupId = groupID,
+                    .mesh = App::singleton<SceneCache>(App::resource_loader())->brick_mesh(),
+                    .transform = scaled,
+                    .color = Color::white(),
+                });
             }
 
             // legs
-            for (int i = 0; i < 3; ++i)
-            {
+            for (int i = 0; i < 3; ++i) {
                 // cone mesh has a source height of 2, stretches from -1 to +1 in Y
                 const float coneHeight = 0.75f * halfWidth;
 
                 const Vec3 meshDirection = {0.0f, 1.0f, 0.0f};
-                Vec3 coneDirection = {};
-                coneDirection[i] = 1.0f;
+                const Vec3 coneDirection = Vec3{}.with_element(i, 1.0f);
 
-                Transform t;
-                t.scale.x = 0.5f * halfWidth;
-                t.scale.y = 0.5f * coneHeight;
-                t.scale.z = 0.5f * halfWidth;
-                t.rotation = xform.rotation * rotation(meshDirection, coneDirection);
-                t.position = xform.position + (t.rotation * ((halfWidth + (0.5f * coneHeight)) * meshDirection));
+                const Quat rot = xform.rotation * rotation(meshDirection, coneDirection);
 
-                Color color = {0.0f, 0.0f, 0.0f, 1.0f};
-                color[i] = 1.0f;
+                appendOut.push_back(DrawableThing{
+                    .id = logicalID,
+                    .groupId = groupID,
+                    .mesh = App::singleton<SceneCache>(App::resource_loader())->cone_mesh(),
+                    .transform = {
+                        .scale = 0.5f * Vec3{halfWidth, coneHeight, halfWidth},
+                        .rotation = rot,
+                        .position = xform.position + (rot * ((halfWidth + (0.5f * coneHeight)) * meshDirection)),
+                    },
+                    .color = Color::black().with_element(i, 1.0f),
+                });
 
-                DrawableThing& legCube = appendOut.emplace_back();
-                legCube.id = logicalID;
-                legCube.groupId = groupID;
-                legCube.mesh = App::singleton<SceneCache>(App::resource_loader())->cone_mesh();
-                legCube.transform = t;
-                legCube.color = color;
-                legCube.flags = SceneDecorationFlags::None;
             }
         }
 
@@ -1363,38 +1349,35 @@ namespace osc::mi
 
         DrawableThing generateBodyElSphere(const Body& bodyEl, const Color& color) const
         {
-            DrawableThing rv;
-            rv.id = bodyEl.getID();
-            rv.groupId = MIIDs::BodyGroup();
-            rv.mesh = m_SphereMesh;
-            rv.transform = SphereMeshToSceneSphereTransform(sphereAtTranslation(bodyEl.getXForm().position));
-            rv.color = color;
-            rv.flags = SceneDecorationFlags::None;
-            return rv;
+            return {
+                .id = bodyEl.getID(),
+                .groupId = MIIDs::BodyGroup(),
+                .mesh = m_SphereMesh,
+                .transform = SphereMeshToSceneSphereTransform(sphereAtTranslation(bodyEl.getXForm().position)),
+                .color = color,
+            };
         }
 
         DrawableThing generateGroundSphere(const Color& color) const
         {
-            DrawableThing rv;
-            rv.id = MIIDs::Ground();
-            rv.groupId = MIIDs::GroundGroup();
-            rv.mesh = m_SphereMesh;
-            rv.transform = SphereMeshToSceneSphereTransform(sphereAtTranslation({0.0f, 0.0f, 0.0f}));
-            rv.color = color;
-            rv.flags = SceneDecorationFlags::None;
-            return rv;
+            return {
+                .id = MIIDs::Ground(),
+                .groupId = MIIDs::GroundGroup(),
+                .mesh = m_SphereMesh,
+                .transform = SphereMeshToSceneSphereTransform(sphereAtTranslation({0.0f, 0.0f, 0.0f})),
+                .color = color,
+            };
         }
 
         DrawableThing generateStationSphere(const StationEl& el, const Color& color) const
         {
-            DrawableThing rv;
-            rv.id = el.getID();
-            rv.groupId = MIIDs::StationGroup();
-            rv.mesh = m_SphereMesh;
-            rv.transform = SphereMeshToSceneSphereTransform(sphereAtTranslation(el.getPos(getModelGraph())));
-            rv.color = color;
-            rv.flags = SceneDecorationFlags::None;
-            return rv;
+            return {
+                .id = el.getID(),
+                .groupId = MIIDs::StationGroup(),
+                .mesh = m_SphereMesh,
+                .transform = SphereMeshToSceneSphereTransform(sphereAtTranslation(el.getPos(getModelGraph()))),
+                .color = color,
+            };
         }
 
         Color RedifyColor(const Color& srcColor) const