+++ src/OpenSimCreator/UI/MeshWarper/MeshWarpingTabInputMeshPanel.h
@@ -267,27 +267,24 @@ namespace osc
             const TPSDocumentLandmarkPair& landmarkPair,
             const std::function<void(SceneDecoration&&)>& decorationConsumer) const
         {
-            const std::optional<Vec3> maybeLocation = GetLocation(landmarkPair, m_DocumentIdentifier);
-            if (!maybeLocation) {
+            const std::optional<Vec3> location = GetLocation(landmarkPair, m_DocumentIdentifier);
+            if (not location) {
                 return;  // no source/destination location for the landmark
             }
 
-            SceneDecoration decoration
-            {
+            SceneDecoration decoration{
                 .mesh = m_State->getLandmarkSphereMesh(),
-                .transform = {.scale = Vec3{m_LandmarkRadius}, .position = *maybeLocation},
+                .transform = {.scale = Vec3{m_LandmarkRadius}, .position = *location},
                 .color = IsFullyPaired(landmarkPair) ? m_State->getPairedLandmarkColor() : m_State->getUnpairedLandmarkColor(),
             };
 
             const TPSDocumentElementID landmarkID{landmarkPair.uid, TPSDocumentElementType::Landmark, m_DocumentIdentifier};
-            if (m_State->isSelected(landmarkID))
-            {
-                decoration.flags |= SceneDecorationFlags::IsSelected;
+            if (m_State->isSelected(landmarkID)) {
+                decoration.flags |= SceneDecorationFlag::RimHighlight0;
             }
-            if (m_State->isHovered(landmarkID))
-            {
+            if (m_State->isHovered(landmarkID)) {
                 decoration.color = to_srgb_colorspace(clamp_to_ldr(multiply_luminance(to_linear_colorspace(decoration.color), 1.2f)));
-                decoration.flags |= SceneDecorationFlags::IsHovered;
+                decoration.flags |= SceneDecorationFlag::RimHighlight1;
             }
 
             decorationConsumer(std::move(decoration));
@@ -309,25 +306,21 @@ namespace osc
             const TPSDocumentNonParticipatingLandmark& npl,
             const std::function<void(SceneDecoration&&)>& decorationConsumer) const
         {
-            SceneDecoration decoration
-            {
+            SceneDecoration decoration{
                 .mesh = m_State->getLandmarkSphereMesh(),
-                .transform =
-                {
+                .transform = {
                     .scale = Vec3{GetNonParticipatingLandmarkScaleFactor()*m_LandmarkRadius},
                     .position = npl.location,
                 },
                 .color = m_State->getNonParticipatingLandmarkColor(),
             };
             const TPSDocumentElementID id{npl.uid, TPSDocumentElementType::NonParticipatingLandmark, m_DocumentIdentifier};
-            if (m_State->isSelected(id))
-            {
-                decoration.flags |= SceneDecorationFlags::IsSelected;
+            if (m_State->isSelected(id)) {
+                decoration.flags |= SceneDecorationFlag::RimHighlight0;
             }
-            if (m_State->isHovered(id))
-            {
+            if (m_State->isHovered(id)) {
                 decoration.color = to_srgb_colorspace(clamp_to_ldr(multiply_luminance(to_linear_colorspace(decoration.color), 1.2f)));
-                decoration.flags |= SceneDecorationFlags::IsHovered;
+                decoration.flags |= SceneDecorationFlag::RimHighlight1;
             }
 
             decorationConsumer(std::move(decoration));