+++ src/oscar/Graphics/Scene/SceneHelpers.h
@@ -82,7 +82,7 @@ namespace osc
         float neck_thickness{};
         float head_thickness{};
         Color color = Color::black();
-        SceneDecorationFlags decoration_flags = SceneDecorationFlags::CastsShadows;
+        SceneDecorationFlags decoration_flags = SceneDecorationFlag::Default;
     };
     void draw_arrow(
         SceneCache&,