+++ src/oscar/Graphics/Scene/SceneRendererParams.h
@@ -8,6 +8,7 @@
 
 namespace osc
 {
+    // the parameters associated with a single call to `SceneRenderer::render`
     struct SceneRendererParams final {
 
         static constexpr Color default_light_color()
@@ -25,31 +26,38 @@ namespace osc
             return {0.0f, -0.001f, 0.0f};
         }
 
-        SceneRendererParams();
-
         friend bool operator==(const SceneRendererParams&, const SceneRendererParams&) = default;
 
-        Vec2i dimensions;
-        AntiAliasingLevel antialiasing_level;
-        bool draw_mesh_normals;
-        bool draw_rims;
-        bool draw_shadows;
-        bool draw_floor;
-        float near_clipping_plane;
-        float far_clipping_plane;
-        Mat4 view_matrix;
-        Mat4 projection_matrix;
-        Vec3 view_pos;
-        Vec3 light_direction;
-        Color light_color;  // ignores alpha
-        float ambient_strength;
-        float diffuse_strength;
-        float specular_strength;
-        float specular_shininess;
-        Color background_color;
-        Color rim_color;
-        Vec2 rim_thickness_in_pixels;
-        Vec3 floor_location;
-        float fixup_scale_factor;
+        // output parameters
+        Vec2i dimensions = {1, 1};
+        AntiAliasingLevel antialiasing_level = AntiAliasingLevel::none();
+
+        // flags
+        bool draw_mesh_normals = false;
+        bool draw_rims = true;
+        bool draw_shadows = true;
+        bool draw_floor = true;
+
+        // camera parameters
+        float near_clipping_plane = 0.1f;
+        float far_clipping_plane = 100.0f;
+        Mat4 view_matrix = identity<Mat4>();
+        Mat4 projection_matrix = identity<Mat4>();
+        Vec3 view_pos = {0.0f, 0.0f, 0.0f};
+
+        // shading parameters
+        Vec3 light_direction = {-0.34f, -0.25f, 0.05f};
+        Color light_color = default_light_color();
+        float ambient_strength = 0.01f;
+        float diffuse_strength = 0.55f;
+        float specular_strength = 0.7f;
+        float specular_shininess = 6.0f;
+        Color background_color = default_background_color();
+        Color rim_color = {0.95f, 0.35f, 0.0f, 1.0f};
+        Vec2 rim_thickness_in_pixels = {1.0f, 1.0f};
+
+        // scene parameters
+        Vec3 floor_location = default_floor_location();
+        float fixup_scale_factor = 1.0f;
     };
 }